Crate writeable

writeable is a utility crate of the ICU4X project.

It includes Writeable, a core trait representing an object that can be written to a sink implementing std::fmt::Write. It is an alternative to std::fmt::Display with the addition of a function indicating the number of bytes to be written.

Writeable improves upon std::fmt::Display in two ways:

More efficient, since the sink can pre-allocate bytes.

Smaller code, since the format machinery can be short-circuited.

Examples

use std::fmt;
use writeable::assert_writeable_eq;
use writeable::LengthHint;
use writeable::Writeable;

struct WelcomeMessage<'s> {
    pub name: &'s str,
}

impl<'s> Writeable for WelcomeMessage<'s> {
    fn write_to<W: fmt::Write + ?Sized>(&self, sink: &mut W) -> fmt::Result {
        sink.write_str("Hello, ")?;
        sink.write_str(self.name)?;
        sink.write_char('!')?;
        Ok(())
    }

    fn writeable_length_hint(&self) -> LengthHint {
        // "Hello, " + '!' + length of name
        LengthHint::exact(8 + self.name.len())
    }
}

let message = WelcomeMessage { name: "Alice" };
assert_writeable_eq!(&message, "Hello, Alice!");

// Types implementing `Writeable` are recommended to also implement `fmt::Display`.
// This can be simply done by redirecting to the `Writeable` implementation:
writeable::impl_display_with_writeable!(WelcomeMessage<'_>);

Macros

assert_writeable_eq

Testing macros for types implementing Writeable. The first argument should be a Writeable, the second argument a string, and the third argument (*_parts_eq only) a list of parts ([(usize, usize, Part)]).

assert_writeable_parts_eq

See assert_writeable_eq.

impl_display_with_writeable

Implements Display for types that implement Writeable.

Structs

LengthHint

A hint to help consumers of Writeable pre-allocate bytes before they call write_to.

Part

Parts are used as annotations for formatted strings. For example, a string like Alice, Bob could assign a NAME part to the substrings Alice and Bob, and a PUNCTUATION part to , . This allows for example to apply styling only to names.

Traits

PartsWrite

A sink that supports annotating parts of the string with Parts.

Writeable

Writeable is an alternative to std::fmt::Display with the addition of a length function.