writeable::adapters

Struct Replace

pub struct Replace<A, B, C> {
    pub source: A,
    pub needle: B,
    pub replacement: C,
}

A Writeable adapter that replaces occurrences of a needle with a replacement.

This adapter performs the replacement in a streaming fashion during write_to, requiring zero allocations.

Examples

use writeable::adapters::Replace;
use writeable::assert_writeable_eq;
use writeable::concat_writeable;

let source = concat_writeable!("I 💖 🦀", " and 🦀 loves me!");
let replace = Replace {
    source,
    needle: "🦀",
    replacement: "Rust",
};

assert_writeable_eq!(replace, "I 💖 Rust and Rust loves me!");

Fields
source: A

The source writeable.

needle: B

The needle to search for.

replacement: C

The replacement writeable.