Macro writeable::assert_writeable_eq

macro_rules! assert_writeable_eq {
    ($actual_writeable : expr, $expected_str : expr $(,) ?) => { ... };
    ($actual_writeable : expr, $expected_str : expr, $($arg : tt) +) => { ... };
}

Testing macro for types implementing Writeable. The first argument should be a string, and the second argument should be a &dyn Writeable.

The macro tests for equality of both string content and verifies the size hint.

Examples

use writeable::Writeable;
use writeable::LengthHint;
use writeable::assert_writeable_eq;
use std::fmt;

struct Demo;
impl Writeable for Demo {
    fn write_to<W: fmt::Write + ?Sized>(&self, sink: &mut W) -> fmt::Result {
        sink.write_str("foo")
    }
    fn write_len(&self) -> LengthHint {
        LengthHint::exact(3)
    }
}

assert_writeable_eq!(&Demo, "foo");
assert_writeable_eq!(&Demo, "foo", "Message: {}", "Hello World");