Crate writeable

writeable is a utility crate of the ICU4X project.

It includes Writeable, a core trait representing an object that can be written to a sink implementing std::fmt::Write. It is an alternative to std::fmt::Display with the addition of a function indicating the number of bytes to be written.

Writeable improves upon std::fmt::Display in two ways:

More efficient, since the sink can pre-allocate bytes.

Smaller code, since the format machinery can be short-circuited.

Types implementing Writeable have a defaulted write_to_string function. If desired, types implementing Writeable can manually implement ToString to wrap write_to_string.

Examples

use writeable::Writeable;
use writeable::LengthHint;
use writeable::assert_writeable_eq;
use std::fmt;

struct WelcomeMessage<'s>{
    pub name: &'s str,
}

impl<'s> Writeable for WelcomeMessage<'s> {
    fn write_to<W: fmt::Write + ?Sized>(&self, sink: &mut W) -> fmt::Result {
        sink.write_str("Hello, ")?;
        sink.write_str(self.name)?;
        sink.write_char('!')?;
        Ok(())
    }

    fn write_len(&self) -> LengthHint {
        // "Hello, " + '!' + length of name
        LengthHint::exact(8 + self.name.len())
    }
}

let message = WelcomeMessage { name: "Alice" };
assert_writeable_eq!(&message, "Hello, Alice!");

Macros

assert_writeable_eq

Testing macros for types implementing Writeable. The first argument should be a Writeable, the second argument a string, and the third argument (*_parts_eq only) a list of parts ([(usize, usize, Part)]).

assert_writeable_parts_eq

Structs

LengthHint

A hint to help consumers of Writeable pre-allocate bytes before they call write_to.

Part

Traits

PartsWrite

A sink that supports annotating parts of the string with Parts.

Writeable

Writeable is an alternative to std::fmt::Display with the addition of a length function.