writeable

Struct LengthHint

#[non_exhaustive]
pub struct LengthHint(pub usize, pub Option<usize>);

A hint to help consumers of Writeable pre-allocate bytes before they call write_to.

This behaves like Iterator::size_hint: it is a tuple where the first element is the lower bound, and the second element is the upper bound. If the upper bound is None either there is no known upper bound, or the upper bound is larger than usize.

LengthHint implements std::ops::{Add, Mul} and similar traits for easy composition. During computation, the lower bound will saturate at usize::MAX, while the upper bound will become None if usize::MAX is exceeded.

Tuple Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

0: usize1: Option<usize>
Implementations

impl LengthHint

pub fn undefined() -> Self

pub fn exact(n: usize) -> Self

write_to will use exactly n bytes.

pub fn at_least(n: usize) -> Self

write_to will use at least n bytes.

pub fn at_most(n: usize) -> Self

write_to will use at most n bytes.

pub fn between(n: usize, m: usize) -> Self

write_to will use between n and m bytes.

pub fn capacity(&self) -> usize

Returns a recommendation for the number of bytes to pre-allocate. If an upper bound exists, this is used, otherwise the lower bound (which might be 0).

Examples

use writeable::Writeable;

fn pre_allocate_string(w: &impl Writeable) -> String {
    String::with_capacity(w.writeable_length_hint().capacity())
}

pub fn is_zero(&self) -> bool

Returns whether the LengthHint indicates that the string is exactly 0 bytes long.